requirements.txt
asgiref==3.7.2
Django==4.1.8
django-extensions==3.2.3
django-filter==23.5
django-js-asset==2.2.0
django-mptt==0.14.0
djangorestframework==3.14.0
drf-yasg==1.21.7
gevent==23.9.1
greenlet==3.0.2
gunicorn==21.2.0
honcho==1.1.0
importlib-metadata==7.0.0
inflection==0.5.1
Markdown==3.5.1
packaging==23.2
psycopg2-binary==2.9.9
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
